requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
astroid==4.0.4
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
chardet==7.3.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cohere==5.21.1
colorama==0.4.6
confection==1.3.3
cryptography==46.0.6
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.8.4
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
emoji==2.15.0
faiss-cpu==1.13.2
fastapi==0.135.2
fastavro==1.12.1
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
installer==0.7.0
instructor==1.14.5
isort==8.0.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numba==0.64.0
numpy==2.4.3
olefile==0.47
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pillow==12.1.1
platformdirs==4.9.4
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pylint==4.0.5
pypdf==6.9.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2026.1.post1
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
scikit-network==0.33.5
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
streamlit==1.55.0
tabulate==0.10.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
types-requests==2.33.0.20260327
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.11.8
unstructured-client==0.42.12
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
webencodings==0.5.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
